23 read value with expiry (#24)
* moved expiry into the set_command.rs

* debug to true

* 21-keys-missing-during-tests: modified config_command_actor_handle to accept set handler

* 21-keys-missing-during-tests: green commit reshuffling expiry loop

* 21-keys-missing-during-tests: clean cargo build

* 21-keys-missing-during-tests: minor cleanup

* read-multiple-strings

* renamed dump.rdb to bkup

* renamed dump.rdb to bkup

* initial fix

* 23-read-value-with-expiry - initial fix 2

* 23-read-value-with-expiry: initial fix 3

* 23-read-value-with-expiry: moar infos

* 23-read-value-with-expiry: fixed expiry FC tag

* info on resizedb

* info on string parsing

* infos in value type parsing

* infos in string parsing

* added length as u32 to next 4 byte parser

* changed the 0xFC xFD parsers

* fewer infos

* explicit as u64/u32

* simple infos

* modified 0xFC parser

* modified 0xFC parser again

* 23-read-value-with-expiry: made aux handle things properly

* 23-read-value-with-expiry: simplified aux handling but code is broken

* 23-read-value-with-expiry: finally fixed incorrect integer parsing

* 23-read-value-with-expiry: infos to debugs

* 23-read-value-with-expiry: switched to proper unix timestamp handling of expiry

* added infos to sleep expiry

* added info for now and duration since epoch

* added info for milliseconds

* added proper conversion for expiry timestamps

* renamed dump.rdb

+++ src/protocol.rs
@@ -37,8 +37,8 @@ pub enum SetCommandSetOption {
 
 #[derive(Debug, Clone, Copy)]
 pub enum SetCommandExpireOption {
-    EX(usize),
-    PX(usize),
+    EX(u32), // seconds
+    PX(u64), // milliseconds
     EXAT(usize),
     PXAT(usize),
     KEEPTTL,
